requirements.txt
fastapi
uvicorn
pydantic
python-multipart
PyMuPDF
pdf2image
Pillow
openai
langchain
numpy
opencv-python
aiofiles
pytesseract
pdf2image

